sgm_cv/requirements.txt
matplotlib==3.6.2
numpy==2.2.6
PyYAML==6.0.2
scipy==1.15.3
seaborn==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
